Improve performance of subscription status methods
The shared header checks the status of the current user's subscription
along with some data from their plan. This performs 5 SQL round trips
just for this data due to how it was being queried from the user model.
It's possible the user may be subscribed personally or via team
membership, and any of those subscriptions may be active or inactive.

This change introduces the concept of a `License` which is a view of
personal and team subscription metadata. The object can answer what a
user has access to and has a corollary `NullSubscription`.

Updating the controller helpers to route through `License` rather than
`current_user` got us down to two queries. The database view cannot
eliminate the second query (for plans) very easily because the data is
represented by a polymorphic association.

In and of itself, this is a small change. The original queries weren't
overly slow, but I found tracing them through the implementation in the
user model to be confusing. They also instantiate a number of ruby
objects to do their work. Ultimately, the license concept could be
pushed forward to eliminate the feature envy methods on user that deal
strictly with `Subscription` and `Plan` data.