Send the user's email in the Cancelled event
When a user cancels, the event (along with others) is propagated to Zapier.

If the event is a cancellation and contains a cancellation reason, we
want Zapier to send an email to our Intercom support address. The body
of the email is the cancellation reason.

This change adds the user's email address to the event payload so that
Zapier knows the user's email address. This allows the email sent to
Intercom to appear to be from the user. This causes Intercom to group
the cancellation reason along with all other messages from the user, and
lets us respond to it as if the user had sent it to us directly.