Allow users to access some videos without a subscription
This commit greatly simplifies how we decide when someone using the site
(subscribers, visitors, and freeloaders (users without an active
subscription)) has access to something.

There were two motivations for this change:

1. We're collapsing our plans down, so that all users have access to
all the site resources.

2. We'd like to offer some videos to freeloaders (they authenticate with
the site and give us their email, we give them some free videos).

This commit establishes a new way of checking for access:

* Subscribers (users with an active subscription) have access to everything.
* Freeloaders (users with no active subscription) have access to a
resource if it responds to `accessible_without_subscription?` with `true`.
* Visitors continue to have access only to previews of things.

A few notable changes in this commit:

* Plans no longer answer questions like `includes_forum?`.
* Subscriptions no longer answer questions about access to things.
* A number of duplicate specs regarding TWI have been collapsed.
* We stopped special-casing videos without a `preview_wistia_id`,
since they all have one (and must, according to validations).