Hide the drip widget on signed_in pages
Recently we made the switch to display the Drip signup form for the Dream Job
course on all pages. While this has improved signups, it also displays the form
for signed in users.

Unfortunately there is no exposed way, either in Semgent or Drip, to configure
the display at this level (signed_in vs visitor).

This change hides the form by adding a simple style rule the `signed_in`
layout.