Only show weekly iteration vids in topic vid list
On the topic pages, we list Trails associated with the given topic first, then
provide a list of individual videos tagged with the topic. This is intended to
be only Weekly Iteration videos, but the existing code was grabbing all videos,
including those from trails.

This double lists the videos, but more problematically, it provides links to
videos that signed out users can't actually see.

This change updates the video display to only show videos from the Weekly
Iteration.