Redirect visitors to sign_in for trail videos
Currently trail videos are listed in the video list on the Topic page. This is
being fixed in another PR, but for now, users are presented with links to these
videos, but the page errors when trying to render them.

This change properly redirects users to `/sign_in` for trail videos, while
still allowing visitors to see the WI preview page.