Fix See What You'll Learn button
The See What You'll Learn button on the subscriptions/new page is
broken for users who are signed in and without an active subscription.
This combination uses the signed_in layout which does not include the
table of contents content used to display the popup.

This PR fixes the button by rendering the table of contents partial as
part of the signed_in layout.