Fire 'Authed to Access' event with video details
In order to put authenticating users into relevant drip segments, we need to
send out an event when they first authenticate. This change adds this event,
"Authed to Access", with the video and watchable (trail or WI) name as
properties.

Initially we tried to track an "Attempted Auth to Access" event when users first
click the button, which would be followed by an actual "Authed to Access" event
after they complete the GitHub auth flow. This would allow us to build a funnel
to analyze how many users bounce in the auth step. Unfortunately we can't do
this truly anonymously and thus we rolled this back / are deferring that work to
a future PR.