Hide inactive beta offers
We've allowed the current set of beta offers to run for a week now and have
received sufficient feedback to move forward. In order to keep the concept of
the beta offers fresh and prevent blindness to them on the practice page, we
want to hide them when not actively seeking feedback.

This changeset allows us to hide beta offers by setting an `active` property to
false.