Use landing layout for sales pages (join, teams)
Currently, the layout is determined exclusively by the user's signed in / out
state. With recent changes to allow users without a subscription to view videos
and more generally interact with the site, we want a bit more subtlety to the
layout logic. Specifically, we always want sales focused pages (/join, /teams)
to be rendered with the simplified layout.

To support the above, this changeset does the following:

- Rename the "visitor" layout to "landing" to better describe its purpose
- Update the layout logic to always show the "landing" layout on /join and
  /teams
- Hide the "Sign In" link for signed in users viewing pages with the "landing"
  layout