Do not link inaccessible content in progress bar
Recently we made it so that in trail step descriptions we only link to
accessible content (user has subscription or the content is accessible without
subscription), but the progress was still rendering links to each step.

This updates the progress bar to only render links with a given dot if the
content is accessible to the current user.

In support of this, the `exercise_for_trail` and `video_for_trail` partials were
combined into one `progress_dot` partial that supports both types of
completable. These partials (and their specs) were nearly identical, and the
interpolation used to render them was difficult to grep for. This change
combines the two into a single partial, and renames to "progress_dot" to better
represent what we're rendering.