Add create_video_markers helper method to irbrc
The current workflow for adding markers to videos is to define them in a local
ruby file, then define this method and run the marker data structure through
that method.

By putting this method in the `.irbrc` it will be available to any console
session, rather than needing to define it for each use.