Update checkout page for conversion optimization
The existing layout and content of the checkout page is very minimal, and has little copy, images, etc that help encourage subscribing. This update revises the copy, introduces an image, updates the sign up button, etc. This is intended as a foundation for further iteration, rather than a final form.