Track sampler vs subscriber on analtyics events
Recently we added the ability for non-subscribers to view the first video or two
for a number of the trails. As a result, we're seeing extremely high viewing
rates for these free sample videos in the analytics which is skewing our
understanding.

This change identifies the user type at event tracking time as either "sampler"
or "subscriber" to allow us to filter.