Update to Rails 4.2.5
This pulled in a major version bump in sprockets due to my local
environment. This caused a breaking test because this new major version
of sprockets defaults to digested assets in all environments. I fixed
the failing test by looking for meta tag content matches that start with
a provided string rather than exact matches. This seemed sufficient for
the test in question.