Pin sprockets-rails to 2.3+ but < 3.0
sprockets-rails 3.0 and sprockets-redirect are incompatible at the
moment. Temporarily pin to sprockets-rails 2.3 until this is fixed. See:
https://github.com/sikachu/sprockets-redirect/issues/12

We use sprockets-redirect for the static error pages (they reference
assets by their raw names). Another potential solution to this would be
to move those error pages into the rails app itself. That could be a
problem for the 500 error page itself depending on the server error.

This was the simplest/quickest fix.