Remove unnecessary parts of `rails_12factor'
rails_12factor is a meta gem which uses `rails_serve_static_assets` and
`rails_stdout_logging` to enable various 12 factor favorable things on
Heroku. We already have the `serve_static_files` setting on in our
application, so we don't need a gem to do it for us. In fact, using an
outdated version of that gem was causing a warning about a deprecated
rails setting.

We still do need the settings enabled by `rails_stdout_logging`, so I
left that gem in place.

This setup is consistent with what suspenders has been doing for some
time.