Improve readability of the video notes
Prior to this change, the notes would span full-width of the main contents
section. This made it difficult to read the notes due to the width.

This change decreases the overall width, and then wraps the notes content in a
box and uses padding to additionally trim down the width while keeping the video
at a reasonable size.