Make CTA button for starting trails dynamic
There's a fairly complex relationship between the current user's role and
whether the trail they're trying to start has a free sample.

This table sums it up:

| role       | has_sample?                     | no_sample              |
| ---------- | --------------------------------|------------------------|
| subscriber | start_trail(first_vid)          | start_trail(first_vid) |
|    sampler | start_trail(trail.sample_video) |            visit trail |
|      guest | auth_to_access                  |            visit trail |

We try to give the user as much as possible; for example, if they're a sampler
and there's a sample video, we show that.