Take users directly to checkout
Why:

Currently we take users to the landing page if they are authenticated
but don't have a subscription and click the "Upcase Membership" link.
This isn't ideal; if they are authenticated then they have been sampling
Upcase content and are presumably sold on subscribing. The landing page
is not for them!

This PR:

* Rewords "Upcase Membership" to "Subscribe".
* This link in the header now takes users directly to the checkout page
  rather then the landing page.