Add Guest - User null object
Why:

Rather then checking for nil on current_user it would be nice to be able
to add more context to that concept. We now have three different user
states, logged_in and subscribed, logged_in and not subscribed and not
logged in.

This PR:

Adds a User null object, Guest, and changes anywhere we're checking
`current_user` to either check `signed_in?` instead.