Subscribe CTA on teams redirect to team checkout
Why:

Currently the teams page has links throughout the page that send to use
to the team checkout. However, the CTA in the header takes the user to
the professional plan checkout. This mismatch in behavior is confusing
and could lead to the user signing up for the wrong plan.

This PR:

* Changes the "Subscribe" CTA in the header to link to the teams
  checkout when on the teams page.
* Changes the teams page to show the landing header no matter what state
  the user is in.
* Adds helpers for linking to the teams checkout and determining if on
  the teams page.