Add tracking events to analytics
Why:

We're about to release trails with more content then ever before! We
currently track when users engage with our video content and find
analysing drop off for those videos across a trail valuable. However, we
currently don't have the same analytics tracking for exercise. This
means that we'll be blind when it comes to analysing how users are
interacting with the exercises contained within a trail.

This PR:

* Adds tracking events for exercise for starting and finishing an
  exercise.
* Adds tracking events for "steps" to track when a video or exercise is
  touched. This will be valuable for combining exercise and video data.
* Adds tracking for starting or finishing a trail.