Unify shows#show page views
Currently the show action decides which view to render based on the
access the current user has to `Show`.

This changes the action to always render `show.html.erb` and use
conditionals to determine what to render in the view using conditionals
testing if the user can access `Show` or not.