Sticky headers for samplers
Why:

We introduced a sticky header for the landing page so that the CTA for
subscribing would always be in the users peripheral. This same logic
should apply to users who are interacting with upcase content but we'd
like to avoid such nonsense once a user is subscribed.

This PR:

* Combines _signed_in_header and _signed_out_header so that the sticky
  logic is consistent if your logged in or not.
* The sticky header is applied unless the users has an active
  subscription.
* Splits out the navigation links depending on if the user is in a
  sales_context or not.
* Changes the sticky header logic to look for a class on the header
  rather then a class on the body.
* Doesn't apply the `viewed-by-visitor` class to the body and update the
  css to reflect this.
* Adds an `encourage_user_to_pay?` helper method for determining if we
  should be applying the sticky header.

https://trello.com/c/sBYD3EuA