Add a script to auto-create markers for videos
Instead of manually creating markers for videos, now we can pipe in a Markdown
file, tell it which Video (by slug) we want to create markers for, and the
script will parse it all out and create the markers for us.

The script expects the piped-in Markdown file to have the markers in a specific
format (documented in the script).

Now that we have a script, we can remove `create_video_markers` from the
`.irbrc`.