Add support for per page title and meta-description
As part of enhancing our relevance in search results we want to clean up our
page titles and meta descriptions as these are core to ranking and to the result
text displayed to searchers.

This change allows for fine grained control of these elements for nearly all
pages, with reasonable fallbacks as needed. The `title` and `<meta
name="Description" />` tags will have one of the following values:

- Explicit `page_title` and `meta_description` field from the model always wins
- Dynamic `page_title` and `meta_description` based on instance's `name`
- Generic site-wide fallback value

This has been implemented for each of  `Topic`, `Trail`, `Video`, and `Show`,
thus covering the vast majority of pages on the site.