Dismiss flash messages when they're clicked on
Previously, the flashes came down from the top of the page and cover most of the
header, making it impossible to click on anything in the header without
reloading the page.

Now flashes still come down from the top of the page, but clicking anywhere on
the flash message hides the flash message.

https://trello.com/c/btUyOWaN