Link to Advanced ActiveRecord Querying in menu
In order to help promote and show off the new course we want to add it to the
"Explore" menu users see when they are on the `/join` page. I'm replacing the
Intermediate Rails as it is a bit out of date and I'd prefer to lead with AARQ
which is up to date and more representative of current content.