Add contact link to account page and footer
We want people to be able to get in touch with us easily, subscribers and
visitors alike.

For most, the Intercom widget will be the main entry point, but some content
blockers will cause it not to show.

This change adds a contact us section to the account page, and shows the Contact
Us link to all users in the footer.