Update Weekly Iteration tiles on topic page
Currently the tiles on the topic pages for the associated Weekly Iteration
episodes are very subtle and don't include the summary at all. We want to
enhance them to make them more welcoming. Specifically:

- Emphasize the teach photos by making them larger
- Include the `summary`, when present, falling back to the truncated notes

This change also refactors and combines a few helpers that were all truncating
and stripping HTML.