Set return_to path to current_path on sign in links
When a user is on a page and signs on, they typically want to be returned to
that page. As an example, when we send out emails to users with links to Weekly
Iteration videos or new trails, many existing users will click through and be in
a signed out state. Currently, when they sign in, they are taken to the
/practice page.

This change appends the `return_to` param to all Sign In links to correspond to
the current_path such that the user will return to that page after signing in.