Update the checkout to emphasize GitHub signin
Currently the majority of users authenticate with GitHub (rather than username &
password auth), but the checkout page is not geared towards this. The GitHub
auth button is one of two equal size and visual emphasis buttons, shown above
the traditional signup form.

With this change, the GitHub auth button is emphasized, and all other fields are
hidden. Users can reveal the more detailed forms if they like, but for the
majority case things are streamlined and more clear.

- Extract some partials to simplify / clean up checkout view
- Hide the forms by default, with links to toggle open
- Emphasize the GitHub button as a CTA