Do not provide download links for preview videos
We're seeing a steady steam of errors in airbrake related to the
`DownloadsController` not being able to find the video. This seems to be due to
users trying to download the preview videos (shown to non-subscribers when
viewing a Weekly Iteration video).

This change removes the download links when showing a preview video. In
addition, this allowed for a nice cleanup of the clip partial removing the need
for a clip and preview_clip as this partial is now only responsible for
displaying the video. Plus, some flexbox FTW!