Update to allow hosting on thoughtbot.com/upcase
Fastly sits in front of thoughtbot.com, primarily acting as a CDN, but it can
also act as a reveres proxy. In this case, we've added a condition to cause
Fastly to proxy all requests to thoughtbot.com/upcase to the Upcase Heroku app,
allowing us to serve the app as a subpath of thoughtbot.com.

To support this, the Upcase application needs to be updated to support being
hosted on a subpath. This change adds that support by doing the following:

- Namespacing all application routes under `/upcase`
- Adding a top level redirect such that upcase.com/* -> thoughtbot.com/upcase/$1
- Updating a number of path configurations to work with the /upcase subpath
- Remove rack canonical host in favor of the rack-rewrite host condition
- Moving all files in public to a /upcase directory
- Use APP_DOMAIN env var more consistently