Restore the download links for videos on mobile
I mistakenly assumed that users were not downloading on mobile and as such I hid
the download links on mobile in ead4c93b, but it sounds like some users were
using the download links on mobile.

This change restores the download links, and also simplifies the DOM and class
structure around the video player.