Redirect to /practice after signing in from /join
Recently we updated the header 'Sign In' link to include a return_to from the
path you clicked it, which then will automatically return users to the page they
were on when they clicked the sign in link.

This works well on content pages like a specific video or the Weekly Iteration
index page, but is likely not desired on the /join or /teams page.

This update will now send users to /practice after signing in on any sales page.