Allow users to reactivate if they cancelled
A user might have clicked cancel and then had a change of heart. We want
it to be simply for them to come back, for them to be reaccpted into the
fold without fear of judgment and without any impediments in their path.
to this end, we have a big button on the Account page that "undoes" the
cancellation and reverts their subscription to what it was before.

https://trello.com/c/8pd5H3uc