Allow users to reactivate if they cancelled
A user might have clicked cancel and then had a change of heart. We want
it to be simply for them to come back, for them to be reaccpted into the
fold without fear of judgment and without any impediments in their path.
to this end, we have a big button on the Account page that "undoes" the
cancellation and reverts their subscription to what it was before.

Also, we want to know how well the Reactivate button is working and how often we
entice someone to come back to us. So in the controllepr we should ping the
analytics with a relevant event.

https://trello.com/c/8pd5H3uc