Allow users to resubscribe easily
This change adds a nice friendly button that tells the users that they can
resubscribe with little to no effort. All the hae to do is click the button and
their card on file will be used. Also there's a link to down the page a spell
where they can change their card info.

https://trello.com/c/8pd5H3uc