Allow users to resubscribe easily
This change adds a nice friendly button that tells the users that they can
resubscribe with little to no effort. All the hae to do is click the button and
their card on file will be used. Also there's a link to down the page a spell
where they can change their card info.

Also, if someone cancels their account and then comes back to us some months
later, we want to know about this. It's a guess that some people just want to
pause and come back later. In order to test this assumption, we need analytics.
That's where this commit comes in.

https://trello.com/c/8pd5H3uc