Move the Sign Out link on the my_account page
It's next to the "Your Information" header. Why is it next to the "Your
Information" header? That makes no sense. Move it to the header, under the
title. People don't need this all the time, so it doesn't need to be all up in
your grill, but it does still need to be visible.

https://trello.com/c/7mKc9Q8W