Allow fake customer to show up in place of missing
So, this one is a little esoteric. When we copy the database from production to
staging (or development) as we often do, we don't copy the Stripe data to the
test environment as well. This makes some sense as we don't wane CC details or
anything else to leak,obviously. But we do have Strips customer ids thate are
trying to be loaded and they're obviously not in the test enviornment.

In this case, we have set it up so that if we know we're in Staging or
Development and we get a special error message from Strips -- it will say that
the customer doesn't exist, but it does in production! -- we'll just come up
with a Stripe customer and give that to the user. Since this is primarily so
that viewing a customer doesn't break, it's not super important what data is
shown.

https://trello.com/c/Qkang98w