Don't include comments on unpublished videos
The Discourse JavaScript will automatically create a thread for any new videos
it encounters (if what does not already exist). While building a trail, we
don't want to have a topic thread as it will indirectly announce the trail
before it's ready.

With this change, we only include the Discourse comments JS when a video is
published (based on the trail or the video itself).