Remove impossible Rails Admin authentication block
In order to determine whether to authenticate or authorize based on the return
value of `current_user`.

If `current_user` is `nil`, then it calls the `authenticate_with` block.

If `current_user` is not `nil`, then it calls the `authorize_with` block.

We have a `Guest` object, so `current_user` is never `nil` -- it's just `Guest`.

Thus, the `authenticate_with` block is never called and can be removed.

This commit also adds a test for a visitor accessing the admin area to confirm
that this block is unnecessary.