Pass `href` argument to `have_link` everywhere
The correct method signature is `have_link(text, href: url)`.

We had a number of `have_link` calls without the `href` argument, like
`have_link(text, url)`.

As far as I can tell, this never worked, because fixing it found a test that was
incorrectly passing. Thankfully, Capybara 2.7 shows a warning when `href` is not
passed.

The `have_search_link` method was never called, so delete it.