Fix favicon display
The layout was still pointing at the unprefixed favicon path. This worked in
production as there is, in fact, a favicon at thougbot.com/favicon.ico, just not
the one we are after.

This change updates the favicon path to include the `/upcase` prefix, allowing
for the correct favicon to be served.