Redirect to thoughtbot.com unless you're fastly
During the domain switch from upcase.com to thoughtbot.com/upcase, we ended up
with the situation where you could go to upcase.com/upcase, which isn't really
helping the SEO situation because we're still on that separate domain. We tried
other configurations but ended up with infinite loops and such. Here, we have
Fastly set a header when it gets the page and the app lets it through because
it's fastly and it needs the page, so we give it the page.

https://trello.com/c/zi3lRFER