Use the old URL when relative videos and comments
The comment system is actually links to threads in the Forum. Since this was
Discourse and was a separate thing, it doesn'y necessarily know about the
domain transition. In order to refer to the same videos that it used to, we
need to massage the URL that we tell the forum that the browser is on so it
serves the right thread.

https://trello.com/c/9DtdIz2J