Un-confuse the upgrade to yearly billing
A few people have complained that they upgraded to yearly billing without
really intending to. We checked the page out, and the styling and behavior of
the page was, in fact, inconsistent with other places in the application where
we charge you money. This page, for example, looked like it was going to take
you to a "checkout" page or something else.

To fix this, we made the button look like other buttons that actually hit your
card and we added a JS confirmation so people have to mash Go twice.

https://trello.com/c/og3fCvEs