Do not show signin link on signin page
The link was rendering with recursively nested `return_to` params, which made it
appear as if there were an infinite number of pages on the site. This introduced
issues while trying to get the site re-indexed under the new domain.

This change hides the signin link on the sign_in page, thus avoiding the
recursive link sadness.