Allow trails to be associated to multiple Topics
This is the first step in migrating from a One-to-many relationship, to a
many-to-many. With this code change in place, we should be able to manually
migrate the data with:

``` ruby
Trail.all.each { |trail| trail.topics << trail.topic } }
```

Following that, we can push another update that removes the database field and
any associated code that expects a `Trail` to have a single `Topic`.