Allow trails to have many topics
We added `has_many :topics` to trails in cb3766e0353189bebe54d37973fb25a7ecdcaf8a, but
did not remove the `belongs_to :topic`. This commit removes the singular topic
and fixes the technical and visual consequences of that change:

* Trail breadcrumbs show every topic (which makes them not-quite-breadcrumbs)
* We show only the first topic's image at on the trail show page (because more
  than one image quickly takes up too much screen space)
* Show all topic labels for each trail

Also, we now use the `topics/_topic` partial to render the names of
`trail.topics`, instead of duplicating that logic.