Re-work how accepting an invitation works
As background, an `invitation` has an associated `email` address that the
invitation is sent to. The user then clicks the link in the email and is taken
to the `acceptances#new` page.

Previously, we always showed people a form that asked for their GitHub username
and Upcase password. If people signed up via GitHub OAuth, they didn't have a
password, and we fielded a lot of confused support requests.

Now we have a new flow:

* User clicks on an emailed invitation link
* If they are signed in when they arrive on the `acceptances#new` page, we
  immediately accept the invitation and redirect to the root path. We don't
  check that `current_user.email == invitation.email` because people have
  personal and work emails, and if they have access to the invitation that's
  good enough.
* If the user is _not_ signed in when they arrive on acceptances#new, we store
  the invitation ID in the session and show the user a GitHub sign in button. We do
  not show a username/password form because we want to strongly encourage
  signing up with GitHub. So the user clicks the GitHub button and is either
  signed in or signed up. Either way, they are then bounced back to
  `auth_callbacks#create`, which then redirects them to `acceptances#new`.
  They're now signed in, so the signed-in flow happens and we accept the
  invitation.

This new flow means we don't need the `Acceptance` model at all. The
`Invitation#accept` method handles everything we need, since we can always get
to a signed-in user.

https://trello.com/c/811zmfIu