Fix and clean up dev:prime task used in bin/setup
Recently we updated Trails to be many to many related to Topics (rather than
Trails belonging to a single Topic) but the dev:prime task did not get updated
completely.

This change fixes the issue with dev:prime, allowing `bin/setup` to run cleanly.
In addition, this removes a few dead methods and explicit references to
`FactoryGirl.create` to clean things up.