Use a GitHub Personal Access Token
The project used the `GITHUB_USER` and `GITHUB_PASSWORD` for
authentication. This interacts poorly with 2FA, but requires that the
user in question has a full Owner role.

Instead, switch to a GitHub Personal Access Token. This provides more
fine-grained control over the scope (such as `write:org`) and separates
general sign in (username, password, 2FA) from API usage.

Closes #1664.