Delete existing markers when updating video notes via script
The `create-notes-and-markers` script scans a set of Markdown, strips
out timestamp markers in headers, and replaces them with jump links that
take the user to that point in the Wistia videos. This change deletes
existing markers to prevent extra (harmless but present) markers from
being created in the database.

The updating of the notes and markers is wrapped in a transaction.

cc: @christoomey