Clean up and simplify welcome page
Let's promote the flashcards another time. Since it is likely that the
person on this page is new and just joined, let's point them to our
more feature-rich areas like Weekly Iteration and Trails. These things
are the reason someone just signed up: they hired Upcase to learn a
topic.

The forum is also not a primary job that people hire Upcase for. It's to
learn a new topic. Let's promote the forum later in a user's journey.

This commit also updates some of the wording on the welcome page.