Respect unsubscribe setting for weekly email
This change adds an additional filter step to ignore users yielded from
`ActiveSubscribers` who have unsubscribed from emails. This was added as an
additional filter, rather than as part of the `ActiveSubscribers` logic.

The thinking was the ActiveSubscribers is responsible for the subscription
state, but the subscribed / unsubscribed from emails is an independent concept.
A user can be in an combination of active/inactive & subscribed/unsubscribed.
Unfortunately we have an overlap in naming related to `subscribed`, but for now
this seems like a workable solution.