Upgrade user to team plan from admin
Currently there are some manual hoops to jump through to get a "normal" user to

1. Create and be part of a team
2. Change their subscription plan, both in the database and stripe, to a team plan

This commit adds a field to the user's edit form in rails_admin that
accepts a new team name. If a team name is given, the right steps will
be performed to upgrade the plan, create the team, etc.