Duplicate some styles for current and new welcome pages
In an effort to separate the marketing pages from the application, we
are going to be temporarily duplicating some code while we isolate
styles. This particular duplication of code is necessary because our
variable naming convention for our current marketing pages a new splash
pages don't play well together. This will be fixed in a later PR.