Clarify instructions for Stripe key setup in development mode
Because the stripe keys are required to set up the team user in the dev prime
script, this script cannot succeed without the .env file having the stripe keys
in the first place.

Previously the setup script was running the `rake dev:prime` task prior to
copying the env file (and thus, prior to a developer modifying their .env file
with the appropriate keys), so the task was failing.