Add mailer previews for current stable of mailers
When working with the transactional emails sent by Upcase there's no way
to view the emails unless you test them being sent from staging or
production. To make it easier to edit AND preview the copy in the emails
these mailer preview objects will give us an index of emails in our
development instance. To view them, visit:

http://localhost:<port>/rails/mailers

Where you'll see a list of the previews implemented in this commit.

* Also delete `app/views/annual_billing_mailer/notification.html.erb` as
  it is no longer being used anywhere in the app.