Transactional emails promote referral program
In a play to nudge people towards using the referral program we can
remind then of its existence by mentioning it in the transactional
emails we send periodically.

This commit adds some unique text to several of those emails (when it
makes sense to) with a link to the "my account" page that automatially
pops up the referral modal window.