Add time to complete info for Intermediate RoR5 (#1758)
We want to test showing completion stats for a single trail
"Intermediate Ruby on Rails(Rails 5). This commit adds times to the
page below the trail title on the index and on individual videos.

The functionality was written generally, so this code introduces some
logic into the video and trail views to only show for this trail. If the
experiment succeeds, we'll remove those view conditionals and
`Video#show_timing?` to roll this out across the site.