Remove performance test for practice show page
Performance doesn't seem to be a problem (see https://trello.com/c/YLUR92VB/259-double-check-performance-for-practiceshow-page)
and the test was broken by our new trail length feature.

(The test didn't check performance directly - it tested that the data
used on the page was being eager loaded. We can't effectively
eager load the length of the videos, and it looks like it's not
important that we do.)