Upgrade migrations class syntax
As of Rails 5, migrations must specify a version when inheriting from
ActiveRecord::Migration. This change updates all existing migrations to
that syntax.

In addition, the schema.rb has changed significantly related to no
longer trying to line up columns and other syntactic changes. The
updated schema.rb is included here as well. It _should_ be functionally
identical to the previous version.