Update The Weekly Iteration CTA button and remove free-tag (#1798)
Updates The Weekly Iteration CTA button from "Subscribe Now" to "Sign in with
GitHub for Free Access" to match rest of CTA buttons.

We also remove a free-tag that showed around some videos if they were
free and marketable. That is not longer needed since all videos will be
free.