Use one universal header
Previously Upcase used separate headers for marketing and for
application pages. This made sense but a lot of content was blocked to
people who weren't subscribed so it was more helpful to send them pretty
sales fluff. Since the content will be available to everyone now - we
think it's more beneficial so show them links to take them to the actual
content regardless of whether or not they're signed in.