Allow signed out users to access the /search page
Note: this change ran into a small issue where the search page was
attempting to track the Search term, and was relying on having a signed
in user with an ID to track against. To solve this I implemented the
`id` method on `Guest` to return the string `"guest"` which I think both
solves the problem, and could potentially be useful from a tracking
perspective.