Change landing links from checkout to practice
Why:
When a user reaches the landing page, we would like to redirect them to
the upcase/practice url and not the sign_up or checkout page. We also
don't want them to see the cta to go to signup on the practice page, for
preference in seeing it only on the content that requires signup.

This PR:
Switches over three links on the landing page to point to the practice
url and removes the cta on the practice page. Relevant tests were
removed and a simple view test for the new landing page links was added.